+++ src/main/java/uk/gov/hmcts/reform/roleassignmentbatch/ApplicationParams.java
@@ -1,10 +1,12 @@
 package uk.gov.hmcts.reform.roleassignmentbatch;
 
-import org.springframework.beans.factory.annotation.Value;
 import static uk.gov.hmcts.reform.roleassignmentbatch.util.Constants.DISABLED;
 import static uk.gov.hmcts.reform.roleassignmentbatch.util.Constants.ENABLED;
+
 import javax.inject.Named;
 import javax.inject.Singleton;
+
+import org.springframework.beans.factory.annotation.Value;
 
 @Named
 @Singleton
