+++ src/test/java/uk/gov/hmcts/reform/roleassignmentbatch/helper/TestDataBuilder.java
@@ -10,7 +10,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.Setter;
-import uk.gov.hmcts.reform.domain.model.CcdCaseUser;
+import uk.gov.hmcts.reform.roleassignmentbatch.domain.model.enums.CcdCaseUser;
 import uk.gov.hmcts.reform.roleassignmentbatch.domain.model.enums.ActorIdType;
 import uk.gov.hmcts.reform.roleassignmentbatch.domain.model.enums.Classification;
 import uk.gov.hmcts.reform.roleassignmentbatch.domain.model.enums.GrantType;
